==> 3-utils-mp4-mp3-processing.ipynb <==
# MP4 and MP3 Processing
This file contains various utility functions to process video and voice files.
- extract mp3 voice from mp4 video

import os
import time
from moviepy.editor import VideoFileClip
from tqdm import tqdm

mp4_engaging_folder = A_PATH

## Test on One File

def extract_audio_from_video(mp4_file, mp3_file):
    # Load the video file
    video = VideoFileClip(mp4_file)

    # Extract the audio
    audio = video.audio

    # Write the audio to an MP3 file
    audio.write_audiofile(mp3_file, codec='mp3')

engaging_mp4_files = os.listdir(mp4_engaging_folder)

engaging_mp3_afile = os.path.splitext(engaging_mp4_files[1])[0]
engaging_mp3_afile

amp4_file = mp4_engaging_folder + "/" + engaging_files[1]
amp4_file

amp3_file = mp4_engaging_folder + "/" + engaging_mp3_afile + ".mp3"
amp3_file

%%time
extract_audio_from_video(amp4_file, amp3_file)

## On Batch Files

mp4_folder = A_PATH

all_files = os.listdir(mp4_folder)
mp4_files = [file for file in all_files if file[-4:] == '.mp4']

sorted(mp4_files)

for afile in sorted(mp4_files):
    print(afile)
    print(os.path.splitext(afile)[0])

for afile in sorted(mp4_files):
    print(afile)
    if os.path.splitext(afile)[1] == '.mp4':
        amp4_file = mp4_folder + "/" + afile
        afile_noext = os.path.splitext(afile)[0]
        amp3_file = mp4_folder + "/" + afile_noext + ".mp3"
        extract_audio_from_video(amp4_file, amp3_file)
